# src/main/java/com/gslab/parivahan/api/PlacesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (1.0.16).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.gslab.parivahan.api;

import java.util.List;

import javax.servlet.http.HttpServletRequest;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

import com.gslab.parivahan.util.RideConstants;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;

@Api(value = "Places")
@RequestMapping("/api")
public interface PlacesApi {

	@ApiOperation(value = "Get All Users", nickname = "Places Autocomplete", notes = "Query for places", response = List.class, responseContainer = "List", tags = {
			"Place Autocomplete", })
	@ApiResponses(value = {
			@ApiResponse(code = 200, message = "OK.", response = List.class, responseContainer = "List"),
			@ApiResponse(code = 400, message = "Bad Request, All parameters are mandatory."),
			@ApiResponse(code = 404, message = "Not Found."),
			@ApiResponse(code = 500, message = "Intenal Sevrer, Unknown Error.") })
	@RequestMapping(value = RideConstants.VERSION_1+"/places", produces = { "application/json" }, method = RequestMethod.GET)
	ResponseEntity<List<String>> searchPlaces(
			@ApiParam(value = "query", required = true, defaultValue = "") @RequestParam(value = "q", required = true) String query,
			@ApiParam(value = "Current location latitude", required = false, defaultValue = "") @RequestParam(value = "lat", required = false) String lat,
			@ApiParam(value = "Current location longtitude", required = false, defaultValue = "") @RequestParam(value = "lng", required = false) String lng,HttpServletRequest request);

}