# swagger.yml
{ 
   "swagger":"2.0",
   "info":{ 
      "description":"Apis for Parivahan Application API",
      "version":"1.0.0",
      "title":"Parivahan Application API",
      "contact":{ 
         "email":""
      },
      "license":{ 
         "name":"Apache 2.0",
         "url":"http://www.apache.org/licenses/LICENSE-2.0.html"
      }
   },
   "host":"localhost",
   "basePath":"/",
...
   "paths":{ 
...
      "/api/v1/places":{ 
         "get":{ 
            "tags":[ 
               "Place Autocomplete"
            ],
            "summary":"Get All Users",
            "description":"Query for places",
            "operationId":"Places Autocomplete",
            "consumes":[ 
               "application/json"
            ],
            "produces":[ 
               "application/json"
            ],
            "parameters":[ 
               { 
                  "name":"Authorization",
                  "in":"header",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"query",
                  "in":"query",
                  "description":"query",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"lat",
                  "in":"query",
                  "description":"lat",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"lng",
                  "in":"query",
                  "description":"lng",
                  "required":false,
                  "type":"string"
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"OK.",
                  "schema":{ 
                     "type":"array",
                     "items":{ 
                        "type":"array",
                        "items":{ 
                           "type":"object"
                        }
                     }
                  }
               },
               "400":{ 
                  "description":"Bad Request, All parameters are mandatory."
               },
               "401":{ 
                  "description":"Unauthorized"
               },
               "403":{ 
                  "description":"Forbidden"
               },
               "404":{ 
                  "description":"Not Found."
               },
               "500":{ 
                  "description":"Intenal Sevrer, Unknown Error."
               }
            }
         }
      },
      "/api/v1/ride":{ 
         "get":{ 
            "tags":[ 
               "developers"
            ],
            "summary":"getAllRides",
            "description":"get All rides",
            "operationId":"getAllRides",
            "consumes":[ 
               "application/json"
            ],
            "produces":[ 
               "application/json"
            ],
            "parameters":[ 
               { 
                  "name":"Authorization",
                  "in":"header",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"Content-Type",
                  "in":"header",
                  "description":"Content-Type",
                  "required":true,
                  "type":"string",
                  "default":"application/json"
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"OK.",
                  "schema":{ 
                     "type":"array",
                     "items":{ 
                        "$ref":"#/definitions/Ride"
                     }
                  }
               },
               "400":{ 
                  "description":"Bad Request."
               },
               "401":{ 
                  "description":"Unauthorized"
               },
               "403":{ 
                  "description":"Forbidden"
               },
               "404":{ 
                  "description":"Not Found."
               },
               "500":{ 
                  "description":"Intenal Sevrer, Unknown Error."
               }
            }
         },
         "post":{ 
            "tags":[ 
               "developers"
            ],
            "summary":"offer ride",
            "description":"Host can offer ride using this api with his vehicle details and Start and end Location",
            "operationId":"offerRide",
            "consumes":[ 
               "application/json"
            ],
            "produces":[ 
               "application/json"
            ],
            "parameters":[ 
               { 
                  "name":"Authorization",
                  "in":"header",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"email",
                  "in":"query",
                  "description":"email",
                  "required":true,
                  "type":"string"
               },
               { 
                  "name":"Content-Type",
                  "in":"header",
                  "description":"Content-Type",
                  "required":true,
                  "type":"string",
                  "default":"application/json"
               },
               { 
                  "in":"body",
                  "name":"rideParams",
                  "description":"Ride Details",
                  "required":true,
                  "schema":{ 
                     "$ref":"#/definitions/Ride"
                  }
               },
               { 
                  "name":"showDetailsOnly",
                  "in":"query",
                  "description":"Show ride details only",
                  "required":false,
                  "type":"boolean",
                  "default":false
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"OK, Ride Added Successfully.",
                  "schema":{ 
                     "$ref":"#/definitions/Ride"
                  }
               },
               "201":{ 
                  "description":"Created"
               },
               "400":{ 
                  "description":"Bad Request, All parameters are mandatory."
               },
               "401":{ 
                  "description":"Unauthorized"
               },
               "403":{ 
                  "description":"Forbidden"
               },
               "404":{ 
                  "description":"Not Found."
               },
               "500":{ 
                  "description":"Intenal Sevrer, Unknown Error"
               }
            }
         },
         "delete":{ 
            "tags":[ 
               "developers"
            ],
            "summary":"Delete offered ride",
            "description":"Host can delete the offered ride using this api",
            "operationId":"DeleteRide",
            "consumes":[ 
               "application/json"
            ],
            "produces":[ 
               "application/json"
            ],
            "parameters":[ 
               { 
                  "name":"Authorization",
                  "in":"header",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"offerCode",
                  "in":"query",
                  "description":"offerCode",
                  "required":false,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"OK, Ride Added Successfully.",
                  "schema":{ 
                     "$ref":"#/definitions/Ride"
                  }
               },
               "204":{ 
                  "description":"No Content"
               },
               "400":{ 
                  "description":"Bad Request, All parameters are mandatory."
               },
               "401":{ 
                  "description":"Unauthorized"
               },
               "403":{ 
                  "description":"Forbidden"
               },
               "404":{ 
                  "description":"Not Found."
               },
               "500":{ 
                  "description":"Intenal Sevrer, Unknown Error"
               }
            }
         }
      },
...
      "/api/v1/shuttle/schedule":{ 
         "post":{ 
            "tags":[ 
               "developers"
            ],
            "summary":"addShuttleSchedule",
            "description":"Add Shuttle Schedule",
            "operationId":"addShuttleSchedule",
            "consumes":[ 
               "application/json"
            ],
            "produces":[ 
               "application/json"
            ],
            "parameters":[ 
               { 
                  "name":"Authorization",
                  "in":"header",
                  "required":false,
                  "type":"string"
               },
               { 
                  "in":"body",
                  "name":"shuttleScheduleVO",
                  "description":"shuttleScheduleVO",
                  "required":true,
                  "schema":{ 
                     "$ref":"#/definitions/ShuttleScheduleVO"
                  }
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"Schedule Added Successfully.",
                  "schema":{ 
                     "$ref":"#/definitions/ShuttleScheduleVO"
                  }
               },
               "201":{ 
                  "description":"Created"
               },
               "400":{ 
                  "description":"Bad Request, All parameters are mandatory."
               },
               "401":{ 
                  "description":"Unauthorized"
               },
               "403":{ 
                  "description":"Forbidden"
               },
               "404":{ 
                  "description":"Not Found."
               },
               "500":{ 
                  "description":"Intenal Sevrer, Unknown Error"
               }
            }
         }
      },
...
   },
   "definitions":{ 
...
      "Host":{ 
         "type":"object",
         "properties":{ 
            "Rating":{ 
               "type":"string",
               "example":"4.3"
            },
            "hostAge":{ 
               "type":"integer",
               "format":"int32",
               "example":26
            },
            "hostGender":{ 
               "type":"string",
               "example":"Male"
            },
            "hostName":{ 
               "type":"string",
               "example":"Swapnil"
            },
            "isGovtIdverified":{ 
               "type":"boolean",
               "example":false
            },
            "isPhoneNumberVerified":{ 
               "type":"boolean",
               "example":true
            },
            "totalHostRatings":{ 
               "type":"integer",
               "format":"int32",
               "example":6
            }
         }
      },
      "LatLng":{ 
         "type":"object",
         "properties":{ 
            "lat":{ 
               "type":"number",
               "format":"double"
            },
            "lng":{ 
               "type":"number",
               "format":"double"
            }
         }
      },
...
      "Ride":{ 
         "type":"object",
         "required":[ 
            "availableSeats",
            "charges",
            "departureDate",
            "departureHours",
            "endCoordinate",
            "startCoordinate"
         ],
         "properties":{ 
            "availableSeats":{ 
               "type":"integer",
               "format":"int32",
               "example":3
            },
            "charges":{ 
               "type":"integer",
               "format":"int32",
               "example":100
            },
            "departureDate":{ 
               "type":"string",
               "example":"2019-02-2T07:46:22:427Z"
            },
            "departureHours":{ 
               "type":"string",
               "example":"0000"
            },
            "encodedPath":{ 
               "type":"string",
               "example":"bgcduh435ry78hjshc984yfhichkidf4yiuwbpxjdskcn"
            },
            "endCoordinate":{ 
               "type":"string",
               "example":"19.2183,72.9781"
            },
            "endLatLng":{ 
               "$ref":"#/definitions/LatLng"
            },
            "endLocation":{ 
               "type":"string",
               "example":"Pune University"
            },
            "host":{ 
               "$ref":"#/definitions/Host"
            },
            "id":{ 
               "type":"string",
               "example":"1200"
            },
            "ridePublishedDate":{ 
               "type":"string",
               "example":"15-08-2018"
            },
            "ridePublishedTime":{ 
               "type":"string",
               "example":"1600"
            },
            "rideRequests":{ 
               "type":"array",
               "items":{ 
                  "$ref":"#/definitions/RideRequestVO"
               }
            },
            "selectedRidePath":{ 
               "type":"string"
            },
            "startCoordinate":{ 
               "type":"string",
               "example":"18.5204,73.8567"
            },
            "startLatLng":{ 
               "$ref":"#/definitions/LatLng"
            },
            "startLocation":{ 
               "type":"string",
               "example":"Gslab pune"
            },
            "stopOvers":{ 
               "type":"array",
               "items":{ 
                  "type":"string"
               }
            },
            "totalRidesOffered":{ 
               "type":"integer",
               "format":"int32",
               "example":18
            },
            "totalSeats":{ 
               "type":"integer",
               "format":"int32",
               "example":4
            },
            "user":{ 
               "$ref":"#/definitions/UserVO"
            },
            "vehicle":{ 
               "type":"string",
               "enum":[ 
                  "car",
                  "bike",
                  "shuttle"
               ]
            }
         }
      },
...
      "RideRequestVO":{ 
         "type":"object",
         "properties":{ 
            "bookingId":{ 
               "type":"string"
            },
            "endLocation":{ 
               "type":"string"
            },
            "ride":{ 
               "$ref":"#/definitions/Ride"
            },
            "rideCharges":{ 
               "type":"integer",
               "format":"int32"
            },
            "rideId":{ 
               "type":"string"
            },
            "startLocation":{ 
               "type":"string"
            },
            "status":{ 
               "type":"string",
               "enum":[ 
                  "ACCEPTED",
                  "DECLINED",
                  "PENDING",
                  "CANCELLED"
               ]
            },
            "user":{ 
               "$ref":"#/definitions/UserVO"
            },
            "userEmail":{ 
               "type":"string"
            }
         }
      },
...
      "ShuttleScheduleVO":{ 
         "type":"object",
         "properties":{ 
            "endLocation":{ 
               "type":"string"
            },
            "schedules":{ 
               "type":"array",
               "items":{ 
                  "type":"string"
               }
            },
            "startLocation":{ 
               "type":"string"
            },
            "totalSeats":{ 
               "type":"integer",
               "format":"int32"
            },
            "user":{ 
               "$ref":"#/definitions/UserVO"
            }
         }
      },
...
      "UserVO":{ 
         "type":"object",
         "properties":{ 
            "email":{ 
               "type":"string"
            },
            "firstName":{ 
               "type":"string"
            },
            "gender":{ 
               "type":"string"
            },
            "lastName":{ 
               "type":"string"
            },
            "mobileNumber":{ 
               "type":"string"
            },
            "name":{ 
               "type":"string"
            },
            "username":{ 
               "type":"string"
            }
         }
      },
...
   }
}